+++ libs/i18n/src/lib/pipes/translate.pipe.ts
@@ -1,0 +1,14 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { CzLabel } from '@cognizone/model-utils';
+import { I18nService } from '../services/i18n.service';
+
+@Pipe({
+  name: 'czTranslate'
+})
+export class TranslatePipe implements PipeTransform {
+  constructor(private i18nService: I18nService) {}
+
+  transform(value: CzLabel, params?: {}, lang?: string): unknown {
+    return this.i18nService.translate(value, params, lang);
+  }
+}