+++ libs/transloco-langstring/src/lib/transloco-langstring.module.ts
@@ -2,6 +2,9 @@ import { NgModule } from '@angular/core';
 
 import { LangStringPipe } from './pipes/lang-string.pipe';
 
+/**
+ * @deprecated please use @cognizone/i18n instead
+ */
 @NgModule({
   declarations: [LangStringPipe],
   imports: [],