+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [2.1.0-beta.4](https://github.com/cognizone/ng-cognizone/compare/v2.1.0-beta.3...v2.1.0-beta.4) (2021-07-27)
+
+
+### Bug Fixes
+
+* deprecating @cognizone/transloco-langstring ([3c60e42](https://github.com/cognizone/ng-cognizone/commit/3c60e42c6293f4e3a0d717e9796e0af7605ece9e))
+
+# [2.1.0-beta.3](https://github.com/cognizone/ng-cognizone/compare/v2.1.0-beta.2...v2.1.0-beta.3) (2021-07-27)
+
+
+### Bug Fixes
+
+* **legi-shared:** renamed legacy theme to classic ([9089b56](https://github.com/cognizone/ng-cognizone/commit/9089b566cb8305ce201bc44972b64e3e4e2684e2))
+
+# [2.1.0-beta.2](https://github.com/cognizone/ng-cognizone/compare/v2.1.0-beta.1...v2.1.0-beta.2) (2021-07-27)
+
+
+### Bug Fixes
+
+* replaced I18nService from legi-shared with the one from the i18n lib ([c230b09](https://github.com/cognizone/ng-cognizone/commit/c230b099af37aa6ef247a18f3592faf5c845f364))
+
+# [2.1.0-beta.1](https://github.com/cognizone/ng-cognizone/compare/v2.0.3...v2.1.0-beta.1) (2021-07-27)
+
+
+### Features
+
+* added i18n and i18n-transloco libraries ([b212fae](https://github.com/cognizone/ng-cognizone/commit/b212fae4c755f622852d4ed0d1269d73be7f7528))
+
 ## [2.0.3](https://github.com/cognizone/ng-cognizone/compare/v2.0.2...v2.0.3) (2021-07-26)